Updated create_symlinks to check existance before backup

+++ create_symlink.sh
@@ -1,7 +1,18 @@
 #!/bin/zsh
-mv ~/.vimrc ~/.dotfiles/backups/
-mv ~/.vim ~/.dotfiles/backups/
+if test -f ~/.vimrc; then
+    echo "Backing up .vimrc to .dotfiles/backups/"
+    mv ~/.vimrc ~/.dotfiles/backups/
+fi
+if test -d ~/.vim; then
+    echo "Backing up .vim to ~/.dotfiles/backups/"
+    mv ~/.vim ~/.dotfiles/backups/
+fi
 ln -s ~/.dotfiles/vim/.vimrc ~/.vimrc
 ln -s ~/.dotfiles/vim ~/.vim
-mv ~/.zshrc ~/.dotfiles/backups
+if test -f ~/.zshrc; then
+    echo "Backing up .zshrc to ~/.dotfiles/backups/"
+    mv ~/.zshrc ~/.dotfiles/backups
+fi
 ln -s ~/.dotfiles/zsh/zshrc ~/.zshrc  
+ls -larth ~/.vim ~/.vimrc ~/.zshrc
+
